process comletion implemented

+++ resources/views/branch/tasks/show.blade.php
@@ -10,7 +10,7 @@
             <div class="row g-2 align-items-center">
                 <div class="col">
                     <h2 class="page-title">
-                        <span class="badge bg-azure">{{$task->code}} - </span>{{$task->title}}
+                        <span class="badge bg-azure mx-1">{{$task->code}}</span> - {{$task->title}}
                     </h2>
                 </div>
             </div>
@@ -20,6 +20,64 @@
     <div class="page-body" x-data="taskData">
         <div class="container-xl">
             <div class="row row-cards">
+                @if($process->status == \App\Helpers\ProcessStatusHelper::REJECTED)
+                    <div class="col-sm-12">
+                        <div class="card">
+                            <div class="card-stamp">
+                                <div class="card-stamp-icon bg-{{$process->getStatusColor()}}"><x-svg.file-alert></x-svg.file-alert></div>
+                            </div>
+                            <div class="card-body">
+                                @csrf
+                                <h3 class="card-title">Topshiriq ijrosi ko'rik uchun topshrilgan</h3>
+                                <p>
+
+                                </p>
+                            </div>
+                        </div>
+                    </div>
+                @endif
+
+                @if($process->status == \App\Helpers\ProcessStatusHelper::PROCESSED)
+                <div class="col-sm-12" x-cloak x-show="uploads && uploads.length">
+                    <div class="card">
+                        <div class="card-stamp">
+                            <div class="card-stamp-icon bg-{{$process->getStatusColor()}}"><x-svg.checkbox></x-svg.checkbox></div>
+                        </div>
+                        <form action="{{route('branch.tasks.complete', $process->id)}}" method="POST" class="card-body">
+                            @csrf
+                            <h3 class="card-title">Topshiriq ijrosini yakunlash</h3>
+                            <p>
+                                Agar barcha kerakli fayllar yukalnganligi va talablar bajarilganiga amin bo'lsangiz topshiriq ijrosini yakunlang.
+                                Yakunlangan topshiriq ma'sul xodimga moderatsiya uchun jo'natiladi va o'zgartirishlar kiritish mumkin bo'lmaydi.
+                                Ijro tekshiruv natijaraliga ko'ra qabul qilinadi yoki qayta to'ldirish uchun bekor qilinadi.
+                                <span class="fw-bolder">Xar br qayta urinish uchun imtiyoz blarriga jarima qo'llaniladi.</span>
+                            </p>
+                            <div class="text-end">
+                                <button class="btn btn-green" onclick="return confirm('Topshiriqni yakunlshni tasdiqlang')">
+                                    <x-svg.checkbox></x-svg.checkbox> Topshiriqni yakunlash
+                                </button>
+                            </div>
+                        </form>
+                    </div>
+                </div>
+                @endif
+                @if($process->status == \App\Helpers\ProcessStatusHelper::COMPLETED)
+                <div class="col-sm-12">
+                    <div class="card">
+                        <div class="card-stamp">
+                            <div class="card-stamp-icon bg-{{$process->getStatusColor()}}"><x-svg.file-search></x-svg.file-search></div>
+                        </div>
+                        <div class="card-body">
+                            @csrf
+                            <h3 class="card-title">Topshiriq ijrosi ko'rik uchun topshrilgan</h3>
+                            <p>
+
+                            </p>
+                        </div>
+                    </div>
+                </div>
+                @endif
+
                 <div class="col-lg-8">
                     <div class="card mb-4">
                         <div class="card-body">
@@ -57,7 +115,7 @@
                 <div class="col-lg-4">
                     <div class="card">
                         <div class="card-body">
-                            <h3 class="card-title">Joriy holat: <span class="ms-2 me-1 badge badge-blink bg-{{$process->getStatusColor()}}"></span> {{$process->getStatusName()}}</h3>
+                            <h3 class="card-title">Joriy holat: <span class="ms-2 me-1 badge badge-blink bg-{{$process->getStatusColor()}}"></span> <span class="badge bg-{{$process->getStatusColor()}}">{{$process->getStatusName()}}</span></h3>
                             <ul class="steps steps-vertical">
                                 @foreach(\App\Helpers\ProcessStatusHelper::STATUSES as $id => $status)
                                     @if(!in_array($id, [\App\Helpers\ProcessStatusHelper::PENDING, \App\Helpers\ProcessStatusHelper::REJECTED]))
@@ -81,6 +139,24 @@
             </div>
             <div class="row row-cards">
 
+                @if($process->editable())
+                    <div class="col-sm-12">
+                        <div class="card">
+                            <div class="card-body">
+                                <h3 class="card-title">Fayllarni yuklash</h3>
+                                <form class="dropzone dz-clickable" id="dropzone-multiple" action="./" autocomplete="off" novalidate="">
+                                    @csrf
+                                    <input type="hidden" name="filable_type" value="{{$process::class}}">
+                                    <input type="hidden" name="filable_id" value="{{$process->id}}">
+                                    <input type="hidden" name="dir" value="{{$process->getUploadDirName()}}">
+                                    <input type="hidden" name="file_name" value="{{$task->code}}">
+
+                                    <div class="dz-default dz-message"><button class="dz-button" type="button">Fayllarni tanlang yoki shu oynaga tashlang</button></div></form>
+                            </div>
+                        </div>
+                    </div>
+                @endif
+
                 <div class="col-md-6 col-sm-12">
                     <div class="card mb-4">
                         <div class="card-body">
@@ -101,9 +177,10 @@
                                     <tr :key="file.id">
                                         <td x-text="file.path"></td>
                                         <td x-text="file.extension" class="text-muted"></td>
-                                        <td x-text="file.size + ' MB'" class="text-muted"></td>
+                                        <td x-text="getFileSize(file.size)" class="text-muted"></td>
                                         <td class="text-muted">
-                                            <button class="btn btn-sm btn-azure">Yuklab olish</button>
+                                            <a :href="fileOpenUrl(file.id)" target="_blank" class="btn btn-sm btn-azure mx-1">Ko'rish</a>
+                                            <a :href="fileLoadUrl(file.id)" target="_blank" class="btn btn-sm btn-azure mx-1">Yuklab olish</a>
                                         </td>
                                     </tr>
                                 </template>
@@ -114,17 +191,9 @@
                 <div class="col-md-6 col-sm-12">
                     <div class="card mb-4">
                         <div class="card-body">
-                            <h3 class="card-title">Fayllarni yuklash</h3>
-                            <form class="dropzone dz-clickable" id="dropzone-multiple" action="./" autocomplete="off" novalidate="">
-                                @csrf
-                                <input type="hidden" name="filable_type" value="{{$process::class}}">
-                                <input type="hidden" name="filable_id" value="{{$process->id}}">
-                                <input type="hidden" name="dir" value="{{$process->getUploadDirName()}}">
-
-                                <div class="dz-default dz-message"><button class="dz-button" type="button">Fayllarni tanlang yoki shu oynaga tashlang</button></div></form>
+                            <h3 class="card-title" x-text="`Ijro uchun yuklangan fayllar (${uploads ? uploads.length : 0})`"></h3>
                         </div>
-
-                        <div class="card-table table-responsive" x-show="uploads && uploads.length" x-cloak>
+                        <div class="card-table table-responsive">
                             <table class="table table-vcenter">
                                 <thead>
                                 <tr>
@@ -135,16 +204,22 @@
                                 </tr>
                                 </thead>
                                 <tbody>
-                                <template x-for="file in uploads">
+                                <template x-for="file in uploads" x-show="uploads && uploads.length" x-cloak>
                                     <tr :key="file.id">
                                         <td x-text="file.path"></td>
                                         <td x-text="file.extension" class="text-muted"></td>
-                                        <td x-text="file.size + ' MB'" class="text-muted"></td>
+                                        <td x-text="getFileSize(file.size)" class="text-muted"></td>
                                         <td class="text-muted">
-                                            <button class="btn btn-sm btn-danger" @click.prevent="removeFile(file.id)">O'chirish</button>
+                                            <a :href="fileOpenUrl(file.id)" target="_blank" class="btn btn-sm btn-azure mx-1">Ko'rish</a>
+                                            @if($process->editable())
+                                                <button class="btn btn-sm btn-danger" @click.prevent="removeFile(file.id)">O'chirish</button>
+                                            @endif
                                         </td>
                                     </tr>
                                 </template>
+                                <tr x-cloak x-show="!uploads || !uploads.length">
+                                    <td colspan="4" class="text-center">Yuklangan fayllar mavjud emas</td>
+                                </tr>
                                 </tbody></table>
                         </div>
                     </div>
